+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+
+### Added
+
+- `OpeningDelayRepository` class for storing the delay before opening.
+- `OpeningDelayCommand` class that uses `OpeningDelayRepository` to implement a opening delay check before next commands.
+- `LimitCommand` now saves limit states.
+
+## Fixed
+
+- `ProduceCommand` now produces items with a unique `Guid`.
+
 ## [1.1.1] - 2023-04-14
 
 ### Added